+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+- 2021.03.03, 04:04 -- v.0.2.13.
+- 2021.03.03, 04:03 -- Minor changes: removed extra `checked` prop from `FormItem`, fixed prop types in `FormActions`.
+- 2021.03.03, 04:02 -- ModalWindow: Used FormActions for actions block (can be specified as buttons list, not react elements).
+- 2021.03.03, 04:00 -- ModalPortal: Removed extra `updateShowWithState` calls after state changes.
+- 2021.03.03, 03:58 -- MenuItem: fixed prop types, added item data parameter (passed to onClick or onAction events), checkedTheme prop.
+- 2021.03.03, 03:56 -- Css config parameters: added `primary` & `secondary` to theme names.
+- 2021.03.03, 03:54 -- Menu: fixed prop types, added chained actions context.
 - 2021.02.26, 14:26 -- PD-3595, v.0.2.12: Typescript support added.
 - 2021.02.25, 21:00 -- PD-3591: Non-blocking popup wrapper (rewrited event handlers mechanism; ugly but working).
 - 2021.02.20, 19:31 -- v.0.2.11.
@@ -101,5 +108,5 @@
 - 2020.10.05, 20:38 -- v.0.0.15: Extracted `demoSupport.jsx` routines, fixture selection from url query (parameter `fixture`), added avaialble fixtures list if fixture to display is not specified, added DemoWrapper & demoTitle exportable ficture parameters for demo rendering.
 
 <!--
- @changed 2021.02.26, 14:27
+ @changed 2021.03.03, 04:04
 -->